test init SFM on ex6 data
the generated 3D point cloud and camera positions are almost identical to ex6 solution code outputs, especially when RANSAC is turned off. Makes me think that the issue may lie in the quality/distribution of detected points.
Plan is to change to more CV2 functions for keypoint detection and description and see what happens